backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pymysql>=1.1.0
aiomysql>=0.2.0
cryptography>=42.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
slowapi>=0.1.9
websockets>=12.0
python-dotenv>=1.0.0
pydantic>=2.0.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.5.0
scipy>=1.13.0
joblib>=1.4.0
lightgbm>=4.0.0
xgboost>=2.0.0
requests>=2.31.0
pytest>=8.0.0
httpx>=0.27.0
